feat: implement bucket versioning resource (and import) and datasource
We also provide terraform files for testing in the
`/tests/terraform/bucket` folder.
As this resource further configures buckets, we append the versioning
resource tests to the regular bucket tests.
Therefore, buckets to test resource imports and datasource reads must
already be created in the StorageGRID before running the tests.
Failures will be expected otherwise.
Name and region for read and imported buckets can be configured via
terraform variables to give testers some flexibility, particularly, in
terms of the buckets' regions.
The bucket versioning resource relies on those names and does not need
to be configured additionally.

This commit also contains traces of linting and formatting in otherwise
completely unrelated files.