refactor: create a dedicated api client for bucket
Particularly with respect to the read functionality of bucket
information, which is shared among the Terraform resource and data
source for buckets, this saves a lot of duplicated code.

To provide a coherent client implementation, we also add methods for
creating, updating and deleting buckets to the dedicated api client.