feat: implement bucket policy
With the bucket policy, we stayed faithful to the general idea of
implementing "additional" bucket resources as individual Terraform
resources if the StorageGrid API permits.
It permits for the bucket policies.

We implement the full set of features that can be implemented using
policies on buckets.

It is currently not supported to specify the policy as JSON and use
the `jsondecode` function to specify the policy.
We foresee different issues with attribute parsing for various use
cases.